test(patina): split no-unsafe-url tests into a separate file to stay under source-length cap
